Update materials & shaders: Now supports multiple light sources. Fix core engine and game loop system. Refactor display class, externalize the input part into its own management class (see InputManager) Add new examples in main.cpp Clean useless assets. Add new gui material with basic gradient background. Update the GuiRenderer class and implement now multiples gui elements. Refactor a bit the light class. Add new loadCubeMap method in the Loader class. Add new profiler class to manage time executions of certains parts of the engine. Add new default SkyboxRenderer class for future works. Update gui Widget class to supports new materials. Update basic shaders to support multiple lights sources. Add new text and gui shaders !

+++ PXL/camera.h
@@ -37,10 +37,10 @@ private:
 	glm::vec2 m_mouse;
 	glm::vec2 m_oldMouse;
 
-	float m_sensitivity;
-	float m_moveSpeed;
-	float m_maxSpeed;
-	float m_delta;
+	double m_sensitivity;
+	double m_moveSpeed;
+	double m_maxSpeed;
+	double m_delta;
 
 	std::map<std::string, int> m_keys;
 	std::map<std::string, bool> m_keys_states;
